pyproject.toml
[project]
name = "auto-trader"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.11.*"
dependencies = [
    "click>=8.2.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
